Fix build for rk3399 (Closes: #956298)
The arm-trusted firmware support for rk3399 requires the compilation of
code for Cortex-M0 MCU which is linked inside the bl31.elf binary at
specific location with some size constraints. The flag -O2 used by
default when compiling M0 code generates a binary slightly bigger,
causing the linker to fail.

An initial approach was to suppress the default -O2 with -O0 to avoid
the error, however this broke other targets. Also it is interesting to
note that Makefile states that varibles such as CFLAGS should not be
propagated.

This patch overrides the default CFLAGS used by debian build systems
to avoid this issue, in the same way LDFLAGS is overridden.

Signed-off-by: Walter Lozano <walter.lozano@collabora.com>