debian/rules: Add sun50i_h6 subplatform without PMIC regulators. (Closes: #1032395)
Commit fb23b104 to TF-A introduced setting up power regulators on the
H6. Unfortunately this broke Ethernet on some H6 boards such as the
Orange Pi One Plus and the Orange Pi 3, because the Ethernet PHY
requires certain regulators to be brought up together, not the way TF-A
does it.

For that reason, commit 67412e4d introduced a build option
SUNXI_SETUP_REGULATORS to be set to 0 for those boards. In that case
TF-A leaves the regulators alone and the kernel sets them up later. It's
documented in docs/plat/allwinner.rst.